Move methods for Pointer to their own source file
