+++ include/caffeine/Memory/Pointer.h
@@ -1,13 +1,12 @@
 #pragma once
 
-#include "caffeine/IR/OperationBase.h"
 #include "caffeine/ADT/SlotMap.h"
+#include "caffeine/IR/OperationBase.h"
 #include "caffeine/Memory/Allocation.h"
 
 namespace caffeine {
 
 class Allocation;
-
 
 using AllocId = typename slot_map<Allocation>::key_type;
 
@@ -90,4 +89,19 @@ public:
   void DebugPrint() const;
 };
 
+inline AllocId Pointer::alloc() const {
+  return alloc_;
 }
+
+inline const OpRef& Pointer::offset() const {
+  return offset_;
+}
+
+inline bool Pointer::is_resolved() const {
+  // TODO: This depends on some internal parts of slot_map which aren't really
+  //       meant to be exposed. It should be fine but if slotmap ever starts
+  //       using a different key type then it'll be necessary to rework this.
+  return alloc_.second != SIZE_MAX;
+}
+
+} // namespace caffeine
